Conflicts:
	fbi_web/src/com/fr/bi/web/less/lib/font.less
	fbi_web/src/com/fr/bi/web/less/utils/font.less
	fbi_web/src/com/fr/bi/web/resources/fonts/iconfont.eot
	fbi_web/src/com/fr/bi/web/resources/fonts/iconfont.svg
	fbi_web/src/com/fr/bi/web/resources/fonts/iconfont.ttf
	fbi_web/src/com/fr/bi/web/resources/fonts/iconfont.woff

+++ fbi_conf/src/com/fr/bi/conf/data/source/operator/add/rowcal/rank/RankDealer.java
@@ -60,7 +60,7 @@ public class RankDealer implements ResultDealer {
 			public void actionPerformed(int row) {
 				Number v = (Number) getter.getValue(row);
 				int rank = map.get(v);
-				travel.actionPerformed(new BIDataValue(row, startCol, rank));
+				travel.actionPerformed(new BIDataValue(row, startCol, (long)rank));
 			}	
 		});
 	}
